requirements.txt
requests
Django==1.11.13
elasticsearch==5.5.3
django_auth_ldap==1.3.0
djangorestframework==3.7.7
djangorestframework_jwt==1.11.0
elasticsearch_dsl==5.4.0
mysqlclient==1.3.12
Markdown==2.6.10
coreapi==2.3.3
